+++ test/test_decoders/gm7x/gm7x.cpp
@@ -8,7 +8,7 @@ static void test_getCrankAngle(void)
   extern decoder_status_t decoderStatus;
   extern volatile unsigned long toothLastToothTime;
   extern volatile unsigned long toothLastMinusOneToothTime;
-  extern volatile uint8_t toothCurrentCount;
+  extern volatile uint16_t toothCurrentCount;
 
   auto decoder = triggerSetup_GM7X();
